Move incremental test to common and ignore JS target inside it
