+++ tests/defer/src/commonTest/kotlin/test/DeferTest.kt
@@ -6,15 +6,20 @@ import com.apollographql.apollo3.api.ApolloResponse
 import com.apollographql.apollo3.api.Error
 import com.apollographql.apollo3.mockserver.MockServer
 import com.apollographql.apollo3.mockserver.enqueueMultipart
+import com.apollographql.apollo3.mpp.Platform
+import com.apollographql.apollo3.mpp.currentTimeMillis
+import com.apollographql.apollo3.mpp.platform
 import com.apollographql.apollo3.testing.runTest
 import com.benasher44.uuid.uuid4
 import defer.WithFragmentSpreadsQuery
 import defer.WithInlineFragmentsQuery
 import defer.fragment.ComputerFields
 import defer.fragment.ScreenFields
+import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.flow.toList
 import kotlin.test.Test
 import kotlin.test.assertEquals
+import kotlin.test.assertTrue
 
 @OptIn(ApolloExperimental::class)
 class DeferTest {
@@ -240,5 +245,35 @@ class DeferTest {
     mockServer.enqueueMultipart(jsonList)
     val actualResponseList = apolloClient.query(query).toFlow().toList()
     assertResponseListEquals(expectedDataList, actualResponseList)
+  }
+
+  @Test
+  fun payloadsAreReceivedIncrementally() = runTest(before = { setUp() }, after = { tearDown() }) {
+    if (platform() == Platform.Js) {
+      // TODO For now chunked is not supported on JS - remove this check when it is
+      return@runTest
+    }
+    val jsonList = listOf(
+        """{"data":{"computers":[{"__typename":"Computer","id":"Computer1"},{"__typename":"Computer","id":"Computer2"}]},"hasNext":true}""",
+        """{"data":{"cpu":"386","year":1993,"screen":{"__typename":"Screen","resolution":"640x480"}},"path":["computers",0],"hasNext":true}""",
+        """{"data":{"cpu":"486","year":1996,"screen":{"__typename":"Screen","resolution":"800x600"}},"path":["computers",1],"hasNext":true}""",
+        """{"data":{"isColor":false},"path":["computers",0,"screen"],"hasNext":true,"label":"a"}""",
+        """{"data":{"isColor":true},"path":["computers",1,"screen"],"hasNext":false,"label":"a"}""",
+    )
+
+    val delay = 200L
+    mockServer.enqueueMultipart(jsonList, chunksDelayMillis = delay)
+
+    val actualDelays = mutableListOf<Long>()
+    var lastEmitTime = currentTimeMillis()
+    apolloClient.query(WithFragmentSpreadsQuery()).toFlow().collect {
+      actualDelays += currentTimeMillis() - lastEmitTime
+      lastEmitTime = currentTimeMillis()
+    }
+    // Last 2 emissions can arrive together, so ignore last element
+    for (d in actualDelays.dropLast(1)) {
+      // Allow a 10% margin for inaccuracies
+      assertTrue(d >= delay / 1.1)
+    }
   }
 }
